Reorganized files and added placeholders
