+++ app/src/main/AndroidManifest.xml
@@ -38,10 +38,10 @@
         </provider>
 
         <activity
-            android:name=".activities.RegisterActivity"
+            android:name=".authentication.RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.LoginActivity"
+            android:name=".authentication.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -50,7 +50,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activities.MainActivity"
+            android:name=".main.MainActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustPan"/>
     </application>
