+++ app/src/main/java/com/example/recipeapp/models/parse/Post.java
@@ -1,4 +1,4 @@
-package com.example.recipeapp.models;
+package com.example.recipeapp.models.parse;
 
 import android.util.Log;
