+++ app/src/main/java/com/example/recipeapp/socialFeed/SocialFeedFragment.java
@@ -1,4 +1,4 @@
-package com.example.recipeapp.fragments;
+package com.example.recipeapp.socialFeed;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -12,10 +12,10 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.recipeapp.adapters.SocialFeedAdapter;
 import com.example.recipeapp.databinding.FragmentSocialFeedBinding;
 import com.example.recipeapp.models.EndlessRecyclerViewScrollListener;
-import com.example.recipeapp.models.Post;
+import com.example.recipeapp.models.parse.Post;
+import com.example.recipeapp.socialFeed.adapters.SocialFeedAdapter;
 import com.parse.ParseQuery;
 
 import java.util.ArrayList;
