+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="activityLoginController"
-            type="com.example.recipeapp.activities.LoginActivity" />
+            type="com.example.recipeapp.authentication.LoginActivity" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -15,14 +15,14 @@
         android:layout_height="match_parent"
         android:background="@color/teal_700"
         android:contentDescription="App logo"
-        tools:context=".activities.LoginActivity">
+        tools:context=".authentication.LoginActivity">
 
         <ImageView
             android:id="@+id/ivAppLogo"
             android:layout_width="100dp"
             android:layout_height="100dp"
             android:layout_marginStart="10dp"
-            android:layout_marginTop="40dp"
+            android:layout_marginTop="50dp"
             android:layout_marginEnd="10dp"
             android:background="@mipmap/ic_launcher_round"
             app:layout_constraintEnd_toEndOf="parent"
@@ -34,7 +34,7 @@
             style="@style/Title"
             android:layout_width="wrap_content"
             android:layout_height="43dp"
-            android:layout_marginTop="40dp"
+            android:layout_marginTop="50dp"
             android:text="Login"
             android:textColor="@color/white"
             android:textSize="36sp"
@@ -49,7 +49,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginStart="50dp"
-            android:layout_marginTop="40dp"
+            android:layout_marginTop="60dp"
             android:layout_marginEnd="50dp"
             android:hint="Username"
             app:layout_constraintEnd_toEndOf="parent"
