+++ app/src/main/res/layout/fragment_barcode_scan.xml
@@ -6,12 +6,12 @@
     <data>
         <variable
             name="fragmentBarcodeController"
-            type="com.example.recipeapp.fragments.BarcodeScanFragment" />
+            type="com.example.recipeapp.inventory.BarcodeScanFragment" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".fragments.BarcodeScanFragment">
+        tools:context=".inventory.BarcodeScanFragment">
 
         <androidx.camera.view.PreviewView
             android:id="@+id/viewFinder"
