+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#D7ECE9"
-    tools:context=".fragments.ProfileFragment">
+    tools:context=".profile.ProfileFragment">
 
     <TextView
         android:id="@+id/tvPostCount"
