+++ app/src/main/res/layout/fragment_social_feed.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.SocialFeedFragment">
+    tools:context=".socialFeed.SocialFeedFragment">
 
     <TextView
         android:id="@+id/tvDiscover"
