+++ app/src/main/res/layout/fragment_upload_post.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="fragmentUploadPostController"
-            type="com.example.recipeapp.fragments.UploadPostFragment" />
+            type="com.example.recipeapp.socialFeed.UploadPostFragment" />
     </data>
 
     <ScrollView
@@ -18,7 +18,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="#D7ECE9"
-            tools:context=".fragments.UploadPostFragment">
+            tools:context=".socialFeed.UploadPostFragment">
 
             <com.google.android.material.textfield.TextInputLayout
                 android:id="@+id/tilTitle"
